# docs/mobile-handoff/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MedicalPlab Product API",
    "description": "Fail-closed MedicalPlab API for pilot and production clients",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v1/anatomy/session/start": {
      "post": {
        "tags": [
          "Generative 3D Anatomy"
        ],
        "summary": "Start Anatomy Session",
        "description": "Start or resume an anatomy learning session.",
        "operationId": "start_anatomy_session_api_v1_anatomy_session_start_post",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-Learner-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Learner-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/anatomy/session/{session_id}": {
      "get": {
        "tags": [
          "Generative 3D Anatomy"
        ],
        "summary": "Get Anatomy Session",
        "description": "Retrieve session state with ownership enforcement.",
        "operationId": "get_anatomy_session_api_v1_anatomy_session__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-Learner-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Learner-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnatomySession"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/anatomy/session/{session_id}/interact": {
      "post": {
        "tags": [
          "Generative 3D Anatomy"
        ],
        "summary": "Interact With Anatomy Session",
        "description": "Advance dialogue or process physical 3D mesh interaction.",
        "operationId": "interact_with_anatomy_session_api_v1_anatomy_session__session_id__interact_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-Learner-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Learner-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractSessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/anatomy/session/{session_id}/challenge": {
      "post": {
        "tags": [
          "Generative 3D Anatomy"
        ],
        "summary": "Submit Anatomy Challenge",
        "description": "Submit learner 3D mesh selection for deterministic challenge evaluation.",
        "operationId": "submit_anatomy_challenge_api_v1_anatomy_session__session_id__challenge_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-Learner-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Learner-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChallengeSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeSubmitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnatomyActionType": {
        "type": "string",
        "enum": [
          "FOCUS_STRUCTURE",
          "HIGHLIGHT_STRUCTURE",
          "ISOLATE_STRUCTURE",
          "SHOW_STRUCTURE",
          "HIDE_STRUCTURE",
          "SHOW_RELATION",
          "SET_STRUCTURE_OPACITY",
          "RESET_SCENE"
        ],
        "title": "AnatomyActionType",
        "description": "Deterministic 3D scene actions recognized by the frontend Scene Controller."
      },
      "AnatomyAgentResponse": {
        "properties": {
          "tutor_message": {
            "type": "string",
            "title": "Tutor Message"
          },
          "scene_actions": {
            "items": {
              "$ref": "#/components/schemas/SceneAction"
            },
            "type": "array",
            "title": "Scene Actions"
          },
          "interaction_request": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InteractionRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "fallback_applied": {
            "type": "boolean",
            "title": "Fallback Applied",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "tutor_message"
        ],
        "title": "AnatomyAgentResponse",
        "description": "Strict schema emitted by the Structured AI Anatomy Tutor Agent."
      },
...
      "AnatomySession": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "learner_id": {
            "type": "string",
            "title": "Learner Id"
          },
          "learning_objective": {
            "type": "string",
            "title": "Learning Objective"
          },
          "lesson_state": {
            "$ref": "#/components/schemas/LessonState",
            "default": "INTRO"
          },
          "current_target_structure_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Target Structure Id"
          },
          "selected_structure_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Selected Structure Ids"
          },
          "hint_level": {
            "type": "integer",
            "title": "Hint Level",
            "default": 0
          },
          "challenge_state": {
            "type": "string",
            "title": "Challenge State",
            "default": "NOT_STARTED"
          },
          "challenge_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChallengeResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          },
          "updated_at": {
            "type": "number",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "learner_id",
          "learning_objective",
          "created_at",
          "updated_at"
        ],
        "title": "AnatomySession",
        "description": "Persistent anatomy learning session."
      },
      "AnatomyStructure": {
        "properties": {
          "structure_id": {
            "type": "string",
            "title": "Structure Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "source_system": {
            "type": "string",
            "title": "Source System",
            "default": "HRA"
          },
          "source_structure_id": {
            "type": "string",
            "title": "Source Structure Id"
          },
          "ontology_id": {
            "type": "string",
            "title": "Ontology Id"
          },
          "mesh_node_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Mesh Node Names"
          },
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Aliases"
          },
          "region": {
            "type": "string",
            "title": "Region",
            "default": "renal"
          },
          "relationships": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Relationships"
          },
          "learning_objectives": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Learning Objectives"
          },
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Refs"
          },
          "fma_crosswalk": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fma Crosswalk"
          }
        },
        "type": "object",
        "required": [
          "structure_id",
          "display_name",
          "source_structure_id",
          "ontology_id",
          "mesh_node_names"
        ],
        "title": "AnatomyStructure",
        "description": "Deterministic anatomical structure metadata grounded in verified ontologies."
      },
...
      "ChallengeResult": {
        "type": "string",
        "enum": [
          "PENDING",
          "CORRECT",
          "INCORRECT"
        ],
        "title": "ChallengeResult",
        "description": "Deterministic outcome of the independent 3D identification challenge."
      },
      "ChallengeSubmitRequest": {
        "properties": {
          "learner_id": {
            "type": "string",
            "title": "Learner Id"
          },
          "selected_structure_id": {
            "type": "string",
            "title": "Selected Structure Id"
          }
        },
        "type": "object",
        "required": [
          "learner_id",
          "selected_structure_id"
        ],
        "title": "ChallengeSubmitRequest"
      },
      "ChallengeSubmitResponse": {
        "properties": {
          "session": {
            "$ref": "#/components/schemas/AnatomySession"
          },
          "is_correct": {
            "type": "boolean",
            "title": "Is Correct"
          },
          "target_structure_id": {
            "type": "string",
            "title": "Target Structure Id"
          },
          "selected_structure_id": {
            "type": "string",
            "title": "Selected Structure Id"
          },
          "tutor_feedback": {
            "type": "string",
            "title": "Tutor Feedback"
          },
          "scene_actions": {
            "items": {
              "$ref": "#/components/schemas/SceneAction"
            },
            "type": "array",
            "title": "Scene Actions"
          }
        },
        "type": "object",
        "required": [
          "session",
          "is_correct",
          "target_structure_id",
          "selected_structure_id",
          "tutor_feedback"
        ],
        "title": "ChallengeSubmitResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InteractSessionRequest": {
        "properties": {
          "learner_id": {
            "type": "string",
            "title": "Learner Id"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "selected_structure_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Structure Id"
          }
        },
        "type": "object",
        "required": [
          "learner_id"
        ],
        "title": "InteractSessionRequest"
      },
      "InteractSessionResponse": {
        "properties": {
          "session": {
            "$ref": "#/components/schemas/AnatomySession"
          },
          "tutor_response": {
            "$ref": "#/components/schemas/AnatomyAgentResponse"
          },
          "deterministic_feedback": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deterministic Feedback"
          }
        },
        "type": "object",
        "required": [
          "session",
          "tutor_response"
        ],
        "title": "InteractSessionResponse"
      },
      "InteractionRequest": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/InteractionRequestType",
            "default": "IDENTIFY_STRUCTURE"
          },
          "target_structure_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Structure Id"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "InteractionRequest",
        "description": "Request from AI tutor for the student to physically interact with the 3D model."
      },
      "InteractionRequestType": {
        "type": "string",
        "enum": [
          "IDENTIFY_STRUCTURE",
          "EXPLORE_SCENE"
        ],
        "title": "InteractionRequestType",
        "description": "Types of physical interactions requested from the student."
      },
...
      "LessonState": {
        "type": "string",
        "enum": [
          "INTRO",
          "GUIDED_VESSELS",
          "GUIDED_IDENTIFICATION",
          "CHALLENGE_READY",
          "CHALLENGE_ACTIVE",
          "COMPLETED"
        ],
        "title": "LessonState",
        "description": "State machine progression for the guided anatomy lesson."
      },
...
      "SceneAction": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/AnatomyActionType"
          },
          "structure_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Structure Id"
          },
          "opacity": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Opacity"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "SceneAction",
        "description": "Validated structured action dispatched to the Three.js Scene Controller."
      },
...
      "StartSessionRequest": {
        "properties": {
          "learner_id": {
            "type": "string",
            "title": "Learner Id"
          },
          "learning_objective": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Learning Objective",
            "default": "RENAL_BLOOD_FLOW_AND_HILUM"
          }
        },
        "type": "object",
        "required": [
          "learner_id"
        ],
        "title": "StartSessionRequest"
      },
      "StartSessionResponse": {
        "properties": {
          "session": {
            "$ref": "#/components/schemas/AnatomySession"
          },
          "initial_tutor_response": {
            "$ref": "#/components/schemas/AnatomyAgentResponse"
          },
          "active_structures": {
            "items": {
              "$ref": "#/components/schemas/AnatomyStructure"
            },
            "type": "array",
            "title": "Active Structures"
          }
        },
        "type": "object",
        "required": [
          "session",
          "initial_tutor_response",
          "active_structures"
        ],
        "title": "StartSessionResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}